Handle failure to read individual feeds
The way feeds are fetched, if one fetch fails, all of them fail. I
noticed this when one of the sites I'm subscribed to went down for a
couple of days and all my feeds stopped updating.

My implemented solution is a simply try-catch block with console
logging. It's suboptimal because the user doesn't know about the
failure, but it's at least better than the current behavior.