Fix outdated package and extension metadata
The author changed their GitHub name, so many of the links are now in
correct. I have adjusted all URLs and similar fields to account for the
renaming.

In addition, when the version was updated in `manifest.json` for the
last release, it was not updated in `package.json`. These numbers are
now in agreement.